피드백 사항 수정
